+++ samples/todoapp/src/main/java/com/examples/todoapp/data/TaskRepo.kt
@@ -1,15 +1,11 @@
 package com.examples.todoapp.data
 
 import com.examples.todoapp.tasks.TaskCompletedEvent
-import com.jakewharton.rxrelay3.BehaviorRelay
-import io.reactivex.rxjava3.core.Observable
-import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
-import io.reactivex.rxjava3.core.Scheduler
-import java.util.concurrent.TimeUnit
+import kotlinx.coroutines.flow.Flow
 
 interface TaskRepo {
 
-    fun tasks(): Observable<List<Task>>
+    fun tasks(): Flow<List<Task>>
 
     fun onTaskCompleted(event: TaskCompletedEvent)
 }