+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - **Breaking**: Remove `ActivityStore.streams`
 - **Breaking**: Moving `FragmentEnvironment` to `FragmentStore`
 - **Breaking**: Remove `BaseFormulaFragment` interface from `formula-android` module.
+- **Breaking**: `ActivityStoreContext` migrated from RxJava to Kotlin coroutines
 
 ## [0.7.1] - June 28, 2022
 - **Breaking**: Rename `FragmentBindingBuilder` to `FragmentStoreBuilder`